Rename Maho\Attributes namespace to Maho\Config
