+++ app/code/core/Mage/Log/Model/Visitor.php
@@ -214,7 +214,7 @@ class Mage_Log_Model_Visitor extends Mage_Core_Model_Abstract
      * @param \Maho\Event\Observer $observer
      * @return  $this
      */
-    #[\Maho\Attributes\Observer('controller_action_predispatch', area: 'frontend')]
+    #[Maho\Config\Observer('controller_action_predispatch', area: 'frontend')]
     public function initByRequest($observer)
     {
         if ($this->_skipRequestLogging || $this->isModuleIgnored($observer)) {
@@ -259,7 +259,7 @@ class Mage_Log_Model_Visitor extends Mage_Core_Model_Abstract
      * @param \Maho\Event\Observer $observer
      * @return  $this
      */
-    #[\Maho\Attributes\Observer('controller_action_postdispatch', area: 'frontend')]
+    #[Maho\Config\Observer('controller_action_postdispatch', area: 'frontend')]
     public function saveByRequest($observer)
     {
         if ($this->_skipRequestLogging || $this->isModuleIgnored($observer)) {
@@ -285,7 +285,7 @@ class Mage_Log_Model_Visitor extends Mage_Core_Model_Abstract
      * @param \Maho\Event\Observer $observer
      * @return  $this
      */
-    #[\Maho\Attributes\Observer('customer_login', area: 'frontend')]
+    #[Maho\Config\Observer('customer_login', area: 'frontend')]
     public function bindCustomerLogin($observer)
     {
         /** @var Mage_Customer_Model_Customer $customer */
@@ -313,7 +313,7 @@ class Mage_Log_Model_Visitor extends Mage_Core_Model_Abstract
      * @param \Maho\Event\Observer $observer
      * @return  $this
      */
-    #[\Maho\Attributes\Observer('customer_logout', area: 'frontend')]
+    #[Maho\Config\Observer('customer_logout', area: 'frontend')]
     public function bindCustomerLogout($observer)
     {
         if ($this->getCustomerId() && $customer = $observer->getEvent()->getCustomer()) {
@@ -326,7 +326,7 @@ class Mage_Log_Model_Visitor extends Mage_Core_Model_Abstract
      * @param \Maho\Event\Observer $observer
      * @return $this
      */
-    #[\Maho\Attributes\Observer('sales_quote_save_after', area: 'frontend')]
+    #[Maho\Config\Observer('sales_quote_save_after', area: 'frontend')]
     public function bindQuoteCreate($observer)
     {
         /** @var Mage_Sales_Model_Quote $quote */
@@ -344,7 +344,7 @@ class Mage_Log_Model_Visitor extends Mage_Core_Model_Abstract
      * @param \Maho\Event\Observer $observer
      * @return $this
      */
-    #[\Maho\Attributes\Observer('checkout_quote_destroy', area: 'frontend')]
+    #[Maho\Config\Observer('checkout_quote_destroy', area: 'frontend')]
     public function bindQuoteDestroy($observer)
     {
         /** @var Mage_Sales_Model_Quote $quote */
